Add sample for packaging and deploying an agent
The sample shows how to deploy an agent as an HTTP API.
The API implements a single endpoint that receives a POST request with
payload containing the state of the environment and returns a response with
the action that the agent takes.

Due to the fact that ``ray`` does not save environment information in the
checkpoints, we need to re-register the sim environment when restoring the
agent.
The information needed are the ``observable_space`` and the ``action_space``.