Remove test_local.py from getting started sample
When ``pytest`` is launched from the root of the repo, it scans all folders
searching for pyton scripts with names starting with the patter ``test_*``.
The file ``test_local.py`` required us to change the CI job for ``pytest``.
With this commit we remove the file. This will allow us to run ``pytest``
from the root of the repo and thus add tests for samples.