Add Baobab to platotk
Baobab is the name of the web API that mimics the behavior of the original
Bonsai API.
Baobab is useful when you have a sim that was integrated with Bonsai, but
that is not easy to integrate with RLlib.
The integration is performed by running Baobab between RLlib and the external
sim.
This is useful for sims such as Anylogic.