Fix restoring error in deploy-agent sample
Recently, the sample stopped working. The terminal output when the error
occurred only mentioned a Rollout worker not healthy without really
specifying the underlying cause.
We were not able to identify the issue, but noticed that using a different
approach for restoring the agent that does not use unpickling works
correctly.
For this reason, we revert back to the restoring approach that registers the
environment before loading the agent.