Deploy agent overriding num rollout workers
Apparently when restoring an Algorithm using ``from_checkpoint``, Ray tries
to initialize the same number of workers used for training.
In order to avoid this, we re-implement the unpickle approach for restoring
the agent where we override the number of rollout workers in the Algorithm
config.
This issue was raised to the RLlib team: https://github.com/ray-project/ray/issues/36761

Also, the number of rollout workers is set to 0 to avoid the bug solved in
071fd69.